Call Leightweight Annotation Widget.

Call Leightweight Annotation Widget.

annotate(
  x,
  annotations = annotationstable(),
  width = NULL,
  height = NULL,
  buttons = list(keep = "yellow", drop = "orange"),
  file = NULL,
  ...
)

Arguments

x
Either a fulltexttable object or an object that can be transformed to a fulltexttable using the fulltexttable() method.

annotations
An optional annotationstable object. If supplied, existing annotations will be highlighted in the fulltext displayed. Defaults to an empty annotationstable generated by calling annotationstable().

width
The width of the annolite htmlwidget.

height
The height of the annolite htmlwidget.

buttons
Named list of length-one character vectors to specify labels of codes and colors that are assigned (names are codes, values colors).

file
If a character vector, a filename to save table with annotations to disk whenever a new annotation is added. If the filename ends with ".rds", a RDS file is saved. In all other cases, a csv file is generated.

...
Further arguments passed into call of fulltexttable.

Value

A data.frame with annotations is returned invisibly.

Examples

library(polmineR)
use("polmineR")

#> ... activating corpus: GERMAPARLMINI (version: 0.0.1 | build date: 2019-02-23)

#> ... activating corpus: REUTERS

P <- partition("GERMAPARLMINI", speaker = "Volker Kauder", date = "2009-11-10")

#> ... get encoding: latin1

#> ... get cpos and strucs

if (interactive()) Y <- annotate(P)
if (interactive()) Y <- annotate(D, annotations = sample_annotation)