Annolite htmlwidget.

Annolite htmlwidget.

annolite(data, width = NULL, height = NULL, dialog = NULL, box = TRUE)

sample_annotation

emma_chapters_tokenized

secretary_general_2000

unga_migrationspeeches_fulltext

unga_migrationspeeches_anntationstable

Arguments

data
The data that is passed to the JavaScript that generates the output. Expected to be a list of lists that provide data on sections of text. Each of the sub-lists is to be a named list of a character vector with the HTML element the section will be wrapped into, and a data.frame (or a list) with a column "token", and a column "id".

width
The width of the widget.

height
The height of the widget.

dialog
The dialog.

box
Logical, whether to put text into a box.

Format

An object of class data.frame with 4 rows and 6 columns.

An object of class list of length 55.

An object of class fulltextlist (inherits from list, vector) of length 21.

An object of class list of length 25.

An object of class annotationstable (inherits from data.frame, list, oldClass, vector) with 1844 rows and 6 columns.

Examples

library(polmineR)
P <- partition("GERMAPARLMINI", speaker = "Volker Kauder", date = "2009-11-10")

#> ... get encoding: latin1

#> ... get cpos and strucs

y <- annolite(
  data = list(paragraphs = fulltextlist(P), annotations = annotationstable()),
  width = "100%",
  dialog = list(choices = dialog_radio_buttons(keep = "yellow", drop = "orange"))
)